Move test only `stripSymbols` utility out of the main bundle
