+++ src/__tests__/utils/stripSymbols.ts
