+++ src/cache/inmemory/__tests__/cache.ts
@@ -1,6 +1,6 @@
 import gql, { disableFragmentWarnings } from 'graphql-tag';
 
-import { stripSymbols } from '../../../utilities/common/stripSymbols';
+import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
 import { cloneDeep } from '../../../utilities/common/cloneDeep';
 import { InMemoryCache, InMemoryCacheConfig } from '../inMemoryCache';
 import { makeReference } from '../helpers';
