+++ src/core/__tests__/QueryManager/multiple-results.ts
@@ -1,7 +1,7 @@
 // externals
 import gql from 'graphql-tag';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
-import { stripSymbols } from '../../../utilities/common/stripSymbols';
+import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
 
 // mocks
 import { MockSubscriptionLink } from '../../../__mocks__/mockLinks';
